% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg13.8.99MgDb.R
\name{gg13.8OTU99MgDb}
\alias{gg13.8OTU99MgDb}
\title{MgDb-class object for greengenes version 13.8 99 OTU}
\description{
gg13.8OTU99MgDb is a MgDb-class object with
sequence and taxonomic data for the greengenes
database version 13.8OTU99, \url{http://greengenes.secondgenome.com}
}
\examples{
gg13.8OTU99MgDb
}
\seealso{
\link[metagenomeFeatures]{MgDb-class}
}

