% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengenes13.8_99MgDb.R
\docType{package}
\name{greengenes13.8OTU99MgDb-package}
\alias{greengenes13.8OTU99MgDb-package}
\title{Greengenes 16S rRNA database MgDb Package.}
\description{
This package loads a MgDb class object with
sequence and taxonomic data for the greengenes
database version 13.8 OTU 99, \url{http://greengenes.secondgenome.com}
}
\examples{
ls('package:greengenes13.8OTU99MgDb')
gg13.8OTU99MgDb
}
\author{
Nathan D. Olson \email{nolson@umiacs.umd.edu}
}
\seealso{
\link[metagenomeFeatures]{MgDb-class}
}
\keyword{package}

